fdisk: use unsigned long long instead int for sectors
The patch is originally based on an old Suse patch.

+++ fdisk/fdisk.h
@@ -81,8 +81,8 @@ extern unsigned int read_int(unsigned int low, unsigned int dflt,
 #define SINGULAR 1
 extern char *const str_units(int);
 
-extern unsigned int get_start_sect(struct partition *p);
-extern unsigned int get_nr_sects(struct partition *p);
+extern unsigned long long get_start_sect(struct partition *p);
+extern unsigned long long get_nr_sects(struct partition *p);
 
 extern int osf_label;
